Restructured webpack files
