+++ package.json
@@ -53,6 +53,7 @@
     "typescript-eslint": "^0.0.1-alpha.0",
     "webpack": "^5.88.1",
     "webpack-cli": "^5.1.4",
+    "webpack-hot-middleware": "^2.25.4",
     "webpack-merge": "^4.2.2"
   },
   "dependencies": {
