+++ webpack/defaultFileTypes.js
@@ -1,0 +1,12 @@
+module.exports = {
+    'application/font': /\.(eot|ttf|woff|woff2)(\?.*)?$/,
+    'application/javascript': /\.(js|jsx)$/,
+    'application/typescript': /\.(ts|tsx)$/,
+    'application/json': /\.json$/,
+    audio: /\.(aac|m4a|mp3|oga|ogg|wav)$/,
+    image: /\.(gif|ico|jpg|jpeg|png|svg|webp)$/i,
+    'text/css': /\.css$/,
+    'text/x-less': /\.less$/,
+    'text/x-sass': /\.(sass|scss)$/,
+    video: /\.(mp4|webm)$/,
+};
